[PR] Add new control UI
It's now possible to use the UI to control per-mount BGM settings without needing to actually get on the mount and use a command. Additionally, there are filtering options in the UI for simplicity. Finally, there's now an option to disable the BGM track "Borderless" (mount default) without needing to manually turn off BGM for the relevant mounts.